Fix error in parsing multi-$ORIGIN zone files
This fixes a bug in parsing zone files that have empty labelled records
after secondary $ORIGIN statements. Records which have an empty label
field should have their label parsed as the same as the previous record.
In files with multiple $ORIGIN statements these empty labels instead
pick up an increasing number of copies of their $ORIGIN value.

This fixes the bug by checking whether the label of the record is the
same as the previous one before determining whether to append the
`unrolled_origin`.